+++ src/EventListener/PrepareAmpTemplateListener.php
@@ -9,9 +9,20 @@
 namespace HeimrichHannot\AmpBundle\EventListener;
 
 use HeimrichHannot\AmpBundle\Event\PrepareAmpTemplateEvent;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 class PrepareAmpTemplateListener
 {
+    /**
+     * @var ContainerInterface
+     */
+    private $container;
+
+    public function __construct(ContainerInterface $container)
+    {
+        $this->container = $container;
+    }
+
     public function onHuhAmpEventPrepareAmpTemplate(PrepareAmpTemplateEvent $event)
     {
         switch ($event->getTemplate()) {
@@ -19,9 +30,18 @@ class PrepareAmpTemplateListener
                 $this->preparePlayerContentElement($event);
 
                 break;
+
+            case 'ce_image':
+
+                break;
         }
     }
 
+    /**
+     * Prepare ce_player for amp.
+     *
+     * @param PrepareAmpTemplateEvent $event
+     */
     protected function preparePlayerContentElement(PrepareAmpTemplateEvent $event)
     {
         $componentsToLoad = $event->getComponentsToLoad();
@@ -32,6 +52,20 @@ class PrepareAmpTemplateListener
             $componentsToLoad[] = 'audio';
         }
         $event->setComponentsToLoad($componentsToLoad);
+        $event->stopPropagation();
+    }
+
+    /**
+     * Prepare ce_image for amp.
+     *
+     * @param PrepareAmpTemplateEvent $event
+     */
+    protected function prepareImageContentElement(PrepareAmpTemplateEvent $event)
+    {
+        $context = $event->getContext();
+        $data = [];
+        $this->container->get('huh.utils.image')->addToTemplateData('singleSRC', 'addImage', $data, $context);
+        $event->setContext($data);
         $event->stopPropagation();
     }
 }