+++ src/Resources/contao/config/config.php
@@ -3,10 +3,10 @@
 /**
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['getPageLayout']['amp'] = ['huh.amp.event_listener.hook_listener', 'getPageLayout'];
-$GLOBALS['TL_HOOKS']['parseTemplate']['amp'] = ['huh.amp.event_listener.hook_listener', 'parseTemplate'];
-$GLOBALS['TL_HOOKS']['generatePage']['amp'] = ['huh.amp.event_listener.hook_listener', 'generatePage'];
-$GLOBALS['TL_HOOKS']['modifyFrontendPage']['amp'] = ['huh.amp.event_listener.hook_listener', 'modifyFrontendPage'];
+$GLOBALS['TL_HOOKS']['getPageLayout']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'getPageLayout'];
+$GLOBALS['TL_HOOKS']['parseTemplate']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'parseTemplate'];
+$GLOBALS['TL_HOOKS']['generatePage']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'generatePage'];
+$GLOBALS['TL_HOOKS']['modifyFrontendPage']['huh_amp'] = ['huh.amp.event_listener.hook_listener', 'modifyFrontendPage'];
 
 /**
  * Frontend modules