+++ src/Resources/contao/dca/tl_layout.php
@@ -52,4 +52,4 @@ $fields = [
     ],
 ];
 
-$dca['fields'] += $fields;
+array_merge($fields, is_array($dca['fields']) ? $dca['fields'] : []);