+++ src/Resources/contao/languages/de/tl_page.php
@@ -7,13 +7,13 @@ $lang = &$GLOBALS['TL_LANG']['tl_page'];
  */
 $lang['encoreEntriesAmp'][0] = 'Aktive Entrys (AMP)';
 $lang['encoreEntriesAmp'][1] = 'Legen Sie hier die aktiven webpack-Entrys für die AMP-Version der Seite fest. Sie können diese Festlegungen auf eventuellen Unterseiten mit Seitenvererbung überschreiben.';
-$lang['amp'][0]              = 'AMP Unterstützung';
-$lang['amp'][1]              = 'AMP Unterstützung aktivieren, deaktivieren oder Feld leer lassen um die Contao Layoutvererbung zu verwenden.';
+$lang['enableAmp'][0]              = 'AMP Unterstützung';
+$lang['enableAmp'][1]              = 'AMP Unterstützung aktivieren, deaktivieren oder Feld leer lassen um die Contao Layoutvererbung zu verwenden.';
 $lang['ampLayout'][0]        = 'AMP Seitenlayout';
 $lang['ampLayout'][1]        = 'Dieses Layout wird verwendet um dem Besucher die AMP-Version der Seite anzuzeigen.';
 
 /**
  * References
  */
-$lang['reference']['amp']['active']   = 'Aktiv';
-$lang['reference']['amp']['inactive'] = 'Inaktiv';
+$lang['reference']['enableAmp']['active']   = 'Aktiv';
+$lang['reference']['enableAmp']['inactive'] = 'Inaktiv';