+++ src/Resources/contao/languages/en/tl_page.php
@@ -7,13 +7,13 @@ $lang = &$GLOBALS['TL_LANG']['tl_page'];
  */
 $lang['encoreEntriesAmp'][0] = 'Active Entries (AMP)';
 $lang['encoreEntriesAmp'][1] = 'Specify here the active webpack entries for the AMP version of the page. You can override these definitions on subpages.';
-$lang['amp'][0]              = 'AMP Support';
-$lang['amp'][1]              = 'Enable or disable AMP support or leave field blank to use Contao layout inheritance.';
+$lang['enableAmp'][0]              = 'AMP Support';
+$lang['enableAmp'][1]              = 'Enable or disable AMP support or leave field blank to use Contao layout inheritance.';
 $lang['ampLayout'][0]        = 'AMP page layout';
 $lang['ampLayout'][1]        = 'This layout is used to show the visitor the AMP version of the page.';
 
 /**
  * References
  */
-$lang['reference']['amp']['active']   = 'Active';
-$lang['reference']['amp']['inactive'] = 'Inactive';
+$lang['reference']['enableAmp']['active']   = 'Active';
+$lang['reference']['enableAmp']['inactive'] = 'Inactive';